Restructure Tests.Core
