+++ src/tests/Tests.Core/PathToolsTests/RemoteUnixPathToolsTests.cs
@@ -1,6 +1,6 @@
 using FileSurfer.Core.Models;
 
-namespace Tests.Core;
+namespace Tests.Core.PathToolsTests;
 
 public class RemoteUnixPathToolsTests
 {
