+++ src/tests/Tests.Core/ResultTests/ResultTests.cs
@@ -1,6 +1,6 @@
 using FileSurfer.Core.Models;
 
-namespace Tests.Core;
+namespace Tests.Core.ResultTests;
 
 public class ResultTests
 {
