+++ src/tests/Tests.Core/ServiceTests/ClipboardManagerTests.cs
@@ -6,7 +6,7 @@ using Mocks;
 using Mocks.Models;
 using Mocks.Services;
 
-namespace Tests.Core;
+namespace Tests.Core.ServiceTests;
 
 public class ClipboardManagerTests
 {
