+++ src/tests/Tests.Core/ServiceTests/DirectoryWatcherTests.cs
@@ -6,7 +6,7 @@ using Mocks.Models;
 
 // ReSharper disable ParameterOnlyUsedForPreconditionCheck.Local
 
-namespace Tests.Core;
+namespace Tests.Core.ServiceTests;
 
 internal sealed class DirectoryWatcherFileInfoProviderMock : MockFileInfoProvider
 {
